Joda Money implementation

+++ src/main/java/ru/savini/fb/controller/AccountController.java
@@ -1,5 +1,6 @@
 package ru.savini.fb.controller;
 
+import java.math.BigDecimal;
 import java.util.List;
 import ru.savini.fb.domain.entity.Account;
 
@@ -7,8 +8,8 @@ public interface AccountController {
     void save(Account account);
     void delete(Account account);
 
-    void putMoney(double moneyAmount, Account account);
-    void withdrawMoney(double moneyAmount, Account account);
+    void putMoney(BigDecimal moneyAmount, Account account);
+    void withdrawMoney(BigDecimal moneyAmount, Account account);
 
     Account getById(long accountId);
     List<Account> getAll();
